pyproject.toml
[project]
name = "bug-sleuth"
version = "0.1.0"
description = "Bug Sleuth Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.20.0",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "aiosqlite>=0.21.0",
    "sqlalchemy>=2.0.41",
    "requests>=2.32.4",
    "httpx>=0.28.1",
    "google-cloud-aiplatform>=1.129.0",
    "google-genai>=1.53.0",
    "opentelemetry-api>=1.37.0",
    "a2a-sdk>=0.1.0",
    "PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


